add one test case in my application
