+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "outdoor-event-fe",
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "tsc & vite",
     "postinstall": "husky install",
@@ -20,11 +21,14 @@
     "@mantine/core": "^7.3.2",
     "@mantine/form": "^7.3.2",
     "@mantine/hooks": "^7.3.2",
+    "@mantine/modals": "^7.3.2",
     "@mantine/notifications": "^7.3.2",
+    "@mantine/styles": "^6.0.21",
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^14.1.2",
     "@types/jest": "^29.5.11",
     "axios": "^1.6.2",
+    "cli-table3": "^0.6.3",
     "dayjs": "^1.11.10",
     "eslint-plugin-storybook": "^0.6.15",
     "i18next": "^23.7.11",
@@ -46,9 +50,11 @@
     "@storybook/addon-links": "^7.6.6",
     "@storybook/addon-onboarding": "^1.0.10",
     "@storybook/blocks": "^7.6.6",
+    "@storybook/jest": "^0.2.3",
     "@storybook/react": "^7.6.6",
     "@storybook/react-vite": "^7.6.6",
     "@storybook/test": "^7.6.6",
+    "@storybook/testing-library": "^0.2.2",
     "@types/react": "^18.2.45",
     "@types/react-dom": "^18.2.18",
     "@typescript-eslint/eslint-plugin": "^6.15.0",
