+++ src/App.tsx
@@ -1,9 +1,19 @@
+import '@mantine/core/styles.css';
+import '@mantine/notifications/styles.css';
+
+import { MantineProvider } from '@mantine/core';
+import { Notifications } from '@mantine/notifications';
 import React from 'react';
 
 import AppRoutes from './router/AppRoutes';
 
 function App() {
-  return <AppRoutes />;
+  return (
+    <MantineProvider>
+      <Notifications />
+      <AppRoutes />
+    </MantineProvider>
+  );
 }
 
 export default App;
