+++ src/main.tsx
@@ -1,8 +1,3 @@
-import '@mantine/core/styles.css';
-import '@mantine/notifications/styles.css';
-
-import { MantineProvider } from '@mantine/core';
-import { Notifications } from '@mantine/notifications';
 import React, { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
@@ -12,9 +7,6 @@ const element = document.getElementById('root');
 const root = createRoot(element!);
 root.render(
   <StrictMode>
-    <MantineProvider>
-      <Notifications />
-      <App />
-    </MantineProvider>
+    <App />
   </StrictMode>,
 );
